requirements-dev.txt
pyresttest
gitpython
